feat: move manifest.json handler to static package and improve path handling
